Probe and sub-class streamlining.
Re-arrange probe initialization / constructors.

Improved consistency in ProcessRecord()s.

More streamlining of generating sample area.

Removed the *2 on sampleArea that was in the 64 diode probes.  Not sure what that was about..

+++ xpms2d/class/define.h
@@ -39,7 +39,7 @@ const size_t MAX_PROBES = 4;
 
 struct recStats
 {
-  int	thisTime;		// Time of this record in secs since midnight.
+  unsigned long	thisTime;	// Time of this record in secs since midnight.
   int	DASelapsedTime;		// milliseconds.
   int	tBarElapsedtime;	// milliseconds.
